tools/zmq_viewer/requirements.txt
pyzmq
opencv-python
numpy
